@@ -7,7 +7,7 @@
     "homepage": "https://github.com/kei-ito/hexo-deployer-aws-s3",
     "repository": "https://github.com/kei-ito/hexo-deployer-aws-s3",
     "engines": {
-        "node": ">=6.9.0"
+        "node": ">=8"
     },
     "main": "lib/index.js",
     "files": [
@@ -20,44 +20,45 @@
         "build:sourceMap": "ts-node test/fixSourceMap.ts",
         "build": "npm-run-all --serial build:cleanup build:tsc build:sourceMap",
         "test:setup": "test/setup.sh",
-        "test:core": "ava -v lib/*.test.js",
-        "test": "npm-run-all --serial build test:setup test:core",
+        "test:core": "ava -v",
+        "test": "npm-run-all --serial test:setup test:core",
         "lint": "eslint --ext .ts src test",
         "version:changelog": "conventional-changelog --preset angular --infile CHANGELOG.md --same-file --release-count 0",
         "version:add": "git add .",
         "version": "npm-run-all --sequential version:changelog version:add"
     },
     "dependencies": {
-        "@types/bluebird": "^3.5.25",
         "@types/glob": "^7.1.1",
         "@types/mime": "^2.0.1",
-        "@types/node": "^11.9.5",
-        "aws-sdk": "^2.411.0",
-        "glob": "^7.1.3",
-        "mime": "^2.4.0",
-        "upath": "^1.1.0"
+        "@types/node": "^12.0.4",
+        "aws-sdk": "^2.468.0",
+        "globby": "^9.2.0",
+        "mime": "^2.4.3",
+        "upath": "^1.1.2"
     },
     "devDependencies": {
-        "@commitlint/cli": "^7.5.2",
-        "@commitlint/config-conventional": "^7.5.0",
-        "@nlib/lint": "^3.3.3",
-        "@types/js-yaml": "^3.12.0",
+        "@commitlint/cli": "^8.0.0",
+        "@commitlint/config-conventional": "^8.0.0",
+        "@nlib/lint": "^3.8.0",
+        "@types/js-yaml": "^3.12.1",
         "@types/mkdirp": "^0.5.2",
-        "@typescript-eslint/eslint-plugin": "^1.4.2",
-        "ava": "^1.2.1",
-        "conventional-changelog-cli": "^2.0.12",
-        "eslint": "^5.14.1",
+        "@typescript-eslint/eslint-plugin": "^1.9.0",
+        "ava": "^2.0.0",
+        "conventional-changelog-cli": "^2.0.21",
+        "eslint": "^5.16.0",
         "hexo": "^3.8.0",
-        "husky": "^1.3.1",
-        "js-yaml": "^3.12.1",
-        "lint-staged": "^8.1.4",
-        "mkdirp": "^0.5.1",
+        "husky": "^2.3.0",
+        "js-yaml": "^3.13.1",
+        "lint-staged": "^8.1.7",
         "npm-run-all": "^4.1.5",
         "rimraf": "^2.6.3",
-        "ts-node": "^8.0.2",
-        "typescript": "^3.3.3333"
+        "ts-node": "^8.2.0",
+        "typescript": "^3.5.1"
     },
     "ava": {
+        "files": [
+            "lib/**/*.test.js"
+        ],
         "compileEnhancements": false,
         "extensions": [
             "ts"
@@ -68,16 +69,29 @@
     },
     "eslintConfig": {
         "extends": "./node_modules/@nlib/lint/.eslintrc.json",
-        "env": {
-            "es6": true,
-            "node": true
-        },
-        "globals": {
-            "hexo": true
-        },
-        "rules": {
-            "no-invalid-this": "off"
-        }
+        "overrides": [
+            {
+                "files": [
+                    "src/**/*.ts",
+                    "test/**/*.ts"
+                ],
+                "env": {
+                    "es6": true,
+                    "node": true
+                }
+            },
+            {
+                "files": [
+                    "src/index.ts"
+                ],
+                "globals": {
+                    "hexo": true
+                },
+                "rules": {
+                    "no-invalid-this": "off"
+                }
+            }
+        ]
     },
     "commitlint": {
         "extends": [
