@@ -1,33 +1,13 @@
 import {EventEmitter} from 'events';
 import {S3 as AWSS3} from 'aws-sdk';
 import {join, dirname} from 'path';
-import $mkdirp = require('mkdirp');
-import {writeFile as $writeFile, createWriteStream} from 'fs';
+import {createWriteStream} from 'fs';
 import {Readable} from 'stream';
+import {mkdirp, writeFile} from './fs';
 
 interface IMyPutObjectOutput {
     promise: () => Promise<AWSS3.Types.PutObjectOutput>,
 }
-
-const mkdirp = (directory: string): Promise<void> => new Promise((resolve, reject) => {
-    $mkdirp(directory, (error) => {
-        if (error) {
-            reject(error);
-        } else {
-            resolve();
-        }
-    });
-});
-
-const writeFile = (file: string, body: string | Buffer): Promise<void> => new Promise((resolve, reject) => {
-    $writeFile(file, body, (error) => {
-        if (error) {
-            reject(error);
-        } else {
-            resolve();
-        }
-    });
-});
 
 const writeFileStream = (file: string, body: Readable): Promise<void> => new Promise((resolve, reject) => {
     body.pipe(createWriteStream(file))
@@ -37,11 +17,14 @@ const writeFileStream = (file: string, body: Readable): Promise<void> => new Pro
 
 export class S3Mock extends EventEmitter {
 
-    public readonly output: string
+    public readonly output: string;
 
-    public readonly config: AWSS3.Types.ClientConfiguration
+    public readonly config: AWSS3.Types.ClientConfiguration;
 
-    public constructor(output: string, config: AWSS3.Types.ClientConfiguration = {}) {
+    public constructor(
+        output: string,
+        config: AWSS3.Types.ClientConfiguration = {},
+    ) {
         super();
         this.output = output;
         this.config = config;
